[FigmaSharp] Fixes GetCodePropertyConfigure and GetViewPropertyConfigure name
